+++ Schedulist.App/Controllers/CalendarController.cs
@@ -1,5 +1,7 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.ModelBinding;
+using Microsoft.AspNetCore.Mvc.Rendering;
+using Schedulist.App.Helper;
 using Schedulist.App.Models;
 using Schedulist.App.Services;
 using Schedulist.DAL;
@@ -150,6 +152,54 @@ namespace Schedulist.App.Controllers
                 Debug.WriteLine($"Exception occurred: {ex.Message}");
                 return View();
             }
+        }
+
+        // GET: WorkModeController/Create
+        public ActionResult CreateWM()
+        {
+            Debug.WriteLine($"Creating Work Mode started!");
+            return View();
+        }
+
+        // POST: WorkModeController/Create
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult CreateWM(WorkModeViewModel workModeView, WorkModesToUser workModesToUser)
+        {
+            var workmodename = WorkModeNamesList.GetAll();
+            workModeView.GetAllWorkModeNames = new List<SelectListItem>();
+            try
+            {
+                if (!ModelState.IsValid)
+                {
+                    return View(workModesToUser);
+                }
+                //foreach (var name in workmodename)
+                //{
+                //    workModeView.GetAllWorkModeNames.Add(new SelectListItem { Text = name.Name, Value = name.Id.ToString() });
+                //}
+                //DateTime selectedDate = (DateTime)TempData.Peek("SelectedDate");
+                //DateOnly parsedChosenDate = DateOnly.FromDateTime(selectedDate);
+                workModeView.WorkModeName = WorkModeNamesList.GetAll().FirstOrDefault(w => w.Id == workModeView.SelectedWorkModeId)?.Name;
+
+                //DateTime selectedDate = (DateTime)TempData.Peek("SelectedDate");
+                //DateOnly parsedChosenDate = DateOnly.FromDateTime(selectedDate);
+
+                WorkModeService _workModeService = new WorkModeService();
+
+                _workModeService.Create(workModeView, workModesToUser);
+                Debug.WriteLine("Created new work mode!");
+                TempData["Success"] = "Work mode has been created successfully";
+                return RedirectToAction(nameof(Index));
+            }
+            catch
+            {
+                return View();
+            }
+
+            //var model = new WorkModeViewModel();
+
+
         }
     }
 }
