+++ Schedulist.App/Models/DayViewModel.cs
@@ -1,4 +1,5 @@
-﻿using Schedulist.DAL;
+﻿using Microsoft.AspNetCore.Mvc.Rendering;
+using Schedulist.DAL;
 
 namespace Schedulist.App.Models
 {
@@ -8,6 +9,8 @@ namespace Schedulist.App.Models
         public User User { get; set; }
         public string WorkMode { get; set; }
         public List<CalendarEvent> CalendarEvents { get; set; }
+        public int SelectedWorkModeId { get; set; }
+        public List<SelectListItem> GetAllWorkModeNames { get; set; }
         public DayViewModel(DateOnly date, User user, string workMode, List<CalendarEvent> calendarEvents)
         {
             Date = date;
