+++ Schedulist.App/Models/WorkModeViewModel.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Schedulist.DAL;
+using System.ComponentModel.DataAnnotations;
 
 namespace Schedulist.App.Models
 {
@@ -9,8 +10,13 @@ namespace Schedulist.App.Models
         public DateOnly Date { get; set; }
         public string WorkModeName { get; set; }
         public int UserID { get; set; }
+        [Display(Name = "Select Work Mode")]
+        public int SelectedWorkModeId { get; set; }
+
         [BindProperty]
         public List<SelectListItem> GetAllWorkModeNames { get;set; }
+
+
