+++ Schedulist.App/Views/Calendar/Day.cshtml
@@ -8,7 +8,19 @@
 <div class="row-margin">
     <h1>@Model.Date</h1>
     <h1>
-        @Model.User.Name @Model.User.Surname - @Model.WorkMode <button class="btn btn-primary" onclick="location.href='@Url.Action("Create", "WorkMode")'">
+        @Model.User.Name @Model.User.Surname - @Model.WorkMode 
+
+@*          @using (Html.BeginForm("Create", "WorkMode", FormMethod.Post))
+        {
+            @Html.AntiForgeryToken()
+
+            @Html.DropDownListFor(model => model.SelectedWorkModeId, Model.GetAllWorkModeNames, "-- Select Work Mode --", new { @class = "form-control" })
+
+            <button type="submit" class="btn btn-primary">
+                Choose work mode
+            </button>
+        } *@
+        <button class="btn btn-primary" onclick="location.href='@Url.Action("CreateWM", "WorkMode")'">
             Choose work mode
         </button>
     </h1>
